+++ qiskit_ibm_runtime/executor_estimator/pec/prepare_pec.py
@@ -15,103 +15,30 @@
 from __future__ import annotations
 
 import logging
-import math
 import sys
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
 
-    from qiskit import QuantumCircuit
     from qiskit.primitives.containers.estimator_pub import EstimatorPub
     from qiskit.quantum_info import PauliLindbladMap
 
-    from ..options_models.measure_noise_learning_options import MeasureNoiseLearningOptions
-    from ..options_models.pec_options import PecOptions
-    from ..options_models.twirling_options import TwirlingOptions
+    from ...options_models.measure_noise_learning_options import MeasureNoiseLearningOptions
+    from ...options_models.pec_options import PecOptions
+    from ...options_models.twirling_options import TwirlingOptions
 
 import numpy as np
-from samplomatic import InjectNoise, build
-from samplomatic.utils import get_annotation
+from samplomatic import build
 
-from ..exceptions import IBMInputValueError
-from ..quantum_program import QuantumProgram
-from ..quantum_program.quantum_program import SamplexItem
-from .trex_utils import create_trex_calibration_circuit
-from .utils import box_circuit, compute_samplex_arguments, make_samplex_arguments
+from ...exceptions import IBMInputValueError
+from ...quantum_program import QuantumProgram
+from ...quantum_program.quantum_program import SamplexItem
+from ..trex_utils import create_trex_calibration_circuit
+from ..utils import box_circuit, compute_samplex_arguments, make_samplex_arguments
+from .pec_utils import calculate_gamma, calculate_pec_twirling_shots
 
 logger = logging.getLogger(__name__)
-
-
-def calculate_gamma(
-    boxed_circuit: QuantumCircuit,
-    noise_model_mapping: dict[str, PauliLindbladMap],
-    noise_factor: float,
-) -> float:
-    """Calculate the PEC gamma factor of a circuit based on a noise model.
-
-    The returned gamma is that associated with the inverse noise maps needed
-    to cancel the noise in the circuit.
-
-    Args:
-        boxed_circuit: The annotated circuit to calculate the PEC gamma for.
-        noise_model_mapping: Mapping between layer ref to a noise model
-        noise_factor: The noise factor of the noise amplification.
-
-    Returns:
-        The PEC gamma factor.
-    """
-    gamma = 1.0
-    for instr in boxed_circuit:
-        if annot := get_annotation(instr.operation, InjectNoise):
-            plm = noise_model_mapping[annot.ref]
-            # scale the noise by noise_factor
-            plm = plm.scale_rates(noise_factor)
-            gamma *= plm.inverse().gamma()
-    return gamma
-
-
-def calculate_pec_twirling_shots(
-    pub_shots: int,
-    num_randomizations: int | str,
-    shots_per_randomization: int | str,
-) -> tuple[int, int]:
-    """Calculate num_randomizations and shots_per_randomization for twirling.
-
-    Implements the logic from TwirlingOptions documentation:
-
-    - If both "auto": shots_per_randomization = 64
-                     num_randomizations = ceil(shots/shots_per_randomization)
-    - If only num_randomizations "auto": num_randomizations = ceil(shots/shots_per_randomization)
-    - If only ``shots_per_randomization`` "auto":
-      shots_per_randomization = ceil(shots/num_randomizations)
-
-    Args:
-        pub_shots: Total shots requested for the pub.
-        num_randomizations: Number of randomizations (or "auto").
-        shots_per_randomization: Shots per randomization (or "auto").
-
-    Returns:
-        Tuple of (num_randomizations, shots_per_randomization).
-    """
-    if num_randomizations == "auto" and shots_per_randomization == "auto":
-        # Both auto: shots_per_rand = max(64, ceil(shots/32))
-        shots_per_rand = 64
-        num_rand = math.ceil(pub_shots / shots_per_rand)
-    elif num_randomizations == "auto":
-        # Only num_rand auto
-        shots_per_rand = int(shots_per_randomization)
-        num_rand = math.ceil(pub_shots / shots_per_rand)
-    elif shots_per_randomization == "auto":
-        # Only shots_per_rand auto
-        num_rand = int(num_randomizations)
-        shots_per_rand = math.ceil(pub_shots / num_rand)
-    else:
-        # Both specified
-        num_rand = int(num_randomizations)
-        shots_per_rand = int(shots_per_randomization)
-
-    return num_rand, shots_per_rand
 
 
 def prepare_pec(
