+++ test/unit/executor_estimator/test_estimator_v2_pea_utils.py
@@ -25,7 +25,7 @@ from samplomatic import InjectNoise
 from samplomatic.utils import get_annotation
 
 from qiskit_ibm_runtime.exceptions import IBMInputValueError
-from qiskit_ibm_runtime.executor_estimator.pea_utils import prepare_pea
+from qiskit_ibm_runtime.executor_estimator.prepare_pea import prepare_pea
 from qiskit_ibm_runtime.executor_estimator.utils import get_layers
 from qiskit_ibm_runtime.options_models.measure_noise_learning_options import (
     MeasureNoiseLearningOptions,
