+++ test/unit/executor_estimator/test_estimator_v2_pec_utils.py
@@ -25,7 +25,8 @@ from samplomatic import InjectNoise
 from samplomatic.utils import get_annotation
 
 from qiskit_ibm_runtime.exceptions import IBMInputValueError
-from qiskit_ibm_runtime.executor_estimator.pec_utils import calculate_gamma, prepare_pec
+from qiskit_ibm_runtime.executor_estimator.pec.pec_utils import calculate_gamma
+from qiskit_ibm_runtime.executor_estimator.pec.prepare_pec import prepare_pec
 from qiskit_ibm_runtime.executor_estimator.utils import get_layers
 from qiskit_ibm_runtime.options_models.measure_noise_learning_options import (
     MeasureNoiseLearningOptions,
